# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  internal_domain_response.ApplyRequest:
    properties:
      message:
        maxLength: 2000
        type: string
      proposed_rate:
        minimum: 0
        type: number
    type: object
  internal_domain_response.ModelSummary:
    properties:
      age:
        type: integer
      city:
        type: string
      gender:
        type: string
      height:
        type: number
      id:
        type: string
      name:
        type: string
      rating:
        type: number
    type: object
  internal_domain_response.ResponseResponse:
    properties:
      accepted_at:
        type: string
      casting_city:
        type: string
      casting_id:
        type: string
      casting_title:
        description: Joined data
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      model:
        allOf:
        - $ref: '#/definitions/internal_domain_response.ModelSummary'
        description: Included when listing responses for a casting
      model_id:
        type: string
      model_name:
        type: string
      proposed_rate:
        type: number
      rejected_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  internal_domain_response.UpdateStatusRequest:
    properties:
      status:
        enum:
        - viewed
        - accepted
        - rejected
        type: string
    required:
    - status
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API server for model agency aggregator.
  termsOfService: http://swagger.io/terms/
  title: MWork API
  version: "1.0"
paths:
...
  /castings/{id}/responses:
    get:
      parameters:
      - description: ID кастинга
        in: path
        name: id
        required: true
        type: string
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_domain_response.ResponseResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Список откликов по кастингу
      tags:
      - Response
    post:
      consumes:
      - application/json
      parameters:
      - description: ID кастинга
        in: path
        name: id
        required: true
        type: string
      - description: Комментарий к отклику
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_domain_response.ApplyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_domain_response.ResponseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Откликнуться на кастинг
      tags:
      - Response
...
  /responses/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID отклика
        in: path
        name: id
        required: true
        type: string
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_response.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_domain_response.ResponseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Обновить статус отклика
      tags:
      - Response
  /responses/my:
    get:
      parameters:
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_domain_response.ResponseResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Мои отклики
      tags:
      - Response
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"